requirements.txt
streamlit
graphviz
fpdf
pillow
pandas
